Removes meta-database functions.
The reason for their removal is that it's almost impossible to codify all
possible database creation options (in fact, prior to their removal, they
didn't include said options at all). Now it's entirely possible that we could
have implemented some sane base-level of options for database creation, but in
addition to the former, semantically speaking, the typeclass should only be
responsible for the data inside of it: that is - its collections. Therefore,
functions that create, get and delete collections inside of a database were
left in, but "meta-database" functions that operate on the database itself were
removed.

It can be argued that Actors can control their own creation, deletion, etc..,
but for the time being, I have decided that `Database` is not a total Actor.