Reconciles impossible function signatures and adds connection functions.
Prior, if you wanted to use `{create,delete}Collection`, you could not because
the type system needed a `(Database datastore)`, but there was no such
constraint on the function. This was fixed by making said function operate on a
`Database datastore`, instead of using Eff magic.

Additionally, reimplement the `{open,close}Database` function (renamed to
`{open,close}Connection`. In a77cec0c2ee595a952b2c16c221583ec22d317fd, it was
determined that it was the responsibility of the user to create and maintain
the database. While this is still true, there should be a unified method for
connecting to and disconnecting from the database. Furthermore, actually having
access to the `Database datastore` allows us to add the missing constraints in
the above-mentioned `{create,delete}Collection`.