Re-works the `Foreign`-based DBObject system into its own DSL, similar to aeson's.
The `Foreign`-based DBObject system was far too brittle, and created cases
where it was very difficult, if not impossible to specify correct type
signatures at times. As an added bonus, not relying on `Foreign`s make it
potentially easier to port between different PureScript backends, as opposed to
just Node's.